# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.DocumentResponse:
    properties:
      checksum_sha256:
        example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        type: string
      created_at:
        example: "2024-01-15T10:00:00Z"
        type: string
      download_url:
        example: http://localhost:8080/api/v1/ns/my-namespace/documents/123e4567-e89b-12d3-a456-426614174000?token=abc.def.123.sig
        type: string
      file_name:
        example: document.pdf
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      title:
        example: document.pdf
        type: string
    type: object
  main.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Document storage and management API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Wayfile Document API
  version: 0.1.0
paths:
  /health:
    get:
      description: Check if the API server is running
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health check
      tags:
      - health
  /ns/{namespace}/documents:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to the specified namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.DocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "409":
          description: File with this content already exists
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Upload a document
      tags:
      - documents
  /ns/{namespace}/documents/{documentID}:
    delete:
      description: Delete a file from the specified namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Document UUID
        in: path
        name: documentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete a document
      tags:
      - documents
    get:
      description: Download a file from the specified namespace
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Document UUID
        in: path
        name: documentID
        required: true
        type: string
      - description: Pre-signed token for authentication
        in: query
        name: token
        type: string
      - description: Token expiration timestamp
        in: query
        name: expires
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "401":
          description: Invalid or expired token
          schema:
            type: string
        "404":
          description: File not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Download a document
      tags:
      - documents
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Document storage operations
  name: documents
- description: Health check endpoints
  name: health